fix: 변경된 스키마맞춰서 router 변경 완료

+++ routes/user.js
@@ -29,12 +29,15 @@ router
       } else {
         res.json({});
       }
-    } catch (e) {
-      console.error(e);
+    } catch (err) {
+      console.error(err);
+      res.status(400).json({ message: 'error' });
     }
   })
   // 회원 가입 및 회원 정보 수정
   .patch(async (req, res) => {
+    if (!req.user)
+      return res.status(400).json({ message: 'no user in session' });
     const id = req.user.id;
     try {
       await User.update(
@@ -48,26 +51,28 @@ router
         },
         { where: { id } },
       );
-      res.status(200).end();
-    } catch (e) {
-      res.status(400).end();
+      res.status(200).json({ message: 'success' });
+    } catch (err) {
+      res.status(400).json({ message: 'error' });
     }
   })
   // 회원 탈퇴
   .delete(async (req, res) => {
+    if (!req.user)
+      return res.status(400).json({ message: 'no user in session' });
     const id = req.user.id;
     try {
       await User.destroy({ where: { id } });
-      res.send('success');
-    } catch (e) {
-      res.send('error');
+      res.status(200).json({ message: 'success' });
+    } catch (err) {
+      res.status(400).json({ message: 'error' });
     }
   });
 
 router.post('/logout', (req, res, next) => {
   try {
     req.logout();
-    res.status(200).end();
+    res.status(200).json({ message: 'success' });
   } catch (err) {
     next(err);
   }
@@ -75,27 +80,31 @@ router.post('/logout', (req, res, next) => {
 
 router.get('/studies', async (req, res) => {
   try {
+    if (!req.user)
+      return res.status(400).json({ message: 'no user in session' });
     const uid = req.user.id;
     const studies = await sequelize.query(
       'SELECT * FROM study WHERE id IN (SELECT studyId FROM studymember WHERE userId=:uid)',
       { replacements: { uid }, type: QueryTypes.SELECT },
     );
     res.json(studies);
-    res.status(200).end();
+    res.status(200).json({ message: 'success' });
   } catch (err) {
     console.error(err);
-    res.status(400).json(err);
+    res.status(400).json({ message: 'error' });
   }
 });
 
 // 사용자 포인트 내역 조회
 router.route('/:id/point-history').get(async (req, res) => {
+  if (!req.user) return res.status(400).json({ message: 'no user in session' });
   const userId = req.user.id;
   try {
     const data = await PointHistory.findAll({ where: { userId } });
-    res.send(data);
-  } catch (e) {
-    res.send('error');
+    res.status(200).json(data);
+  } catch (err) {
+    console.error(err);
+    res.status(400).json({ message: 'error' });
   }
 });
